pyproject.toml
[project]
name = "telegram-video-mp3-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.7",
    "yt-dlp>=2026.3.17",
]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.9",
]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.setuptools.packages.find]
where = ["src"]
